backend/requirements.txt
# Django and DRF
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Authentication
djangorestframework-simplejwt==5.3.1
django-rest-passwordreset==1.4.1

# Database
psycopg2-binary==2.9.10  # For PostgreSQL
# sqlite3 is built into Python for development

# Additional utilities
python-decouple==3.8
drf-yasg==1.21.7  # For API documentation

# Development
django-extensions==3.2.3
werkzeug==3.0.6
